+++ backend/vehicles/src/modules/payment/entities/permit-transaction.entity.ts
