+++ backend/vehicles/src/modules/payment/payment.controller.ts
@@ -18,7 +18,7 @@ import {
 import { ExceptionDto } from '../../common/exception/exception.dto';
 import { PaymentService } from './payment.service';
 import { Public } from 'src/common/decorator/public.decorator';
-import { MotiPayDetailsDto } from './dto/response/read-motiPayUrl.dto';
+import { MotiPayDetailsDto } from './dto/response/read-moti-pay-details.dto';
 import { CreateTransactionDto } from './dto/request/create-transaction.dto';
 import { ReadTransactionDto } from './dto/response/read-transaction.dto';
 import { IUserJWT } from 'src/common/interface/user-jwt.interface';
