+++ backend/vehicles/src/modules/payment/payment.service.ts
@@ -1,6 +1,6 @@
 import { HttpException, Injectable } from '@nestjs/common';
 import * as CryptoJS from 'crypto-js';
-import { IPayment } from './interface/payment.interface';
+import { IPayment } from '../../common/interface/payment.interface';
 import { CreateTransactionDto } from './dto/request/create-transaction.dto';
 import { ReadTransactionDto } from './dto/response/read-transaction.dto';
 import { InjectMapper } from '@automapper/nestjs';
@@ -8,9 +8,9 @@ import { Mapper } from '@automapper/core';
 import { Transaction } from './entities/transaction.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { PermitTransaction } from './entities/permitTransaction.entity';
-import { ReadPermitTransactionDto } from './dto/response/read-permitTransaction.dto';
-import { MotiPayDetailsDto } from './dto/response/read-motiPayUrl.dto';
+import { PermitTransaction } from './entities/permit-transaction.entity';
+import { ReadPermitTransactionDto } from './dto/response/read-permit-transaction.dto';
+import { MotiPayDetailsDto } from './dto/response/read-moti-pay-details.dto';
 import { HttpService } from '@nestjs/axios';
 import { lastValueFrom } from 'rxjs';
 import { ExceptionDto } from 'src/common/exception/exception.dto';
